pyproject.toml
[project]
name = "payments-system"
version = "0.1.0"
description = ""
authors = [
    {name = "KamranYsupov",email = "kamrankek48@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django (==4.2.17)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "loguru (==0.6)",
]

[tool.poetry]
packages = [{include = "payments_system", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
